fix: shared nav panel shell, close drawer on desktop resize, shared month label
