+++ clients/docs/src/app/docs/_components/docs-nav-context.tsx
@@ -65,6 +65,20 @@ export function DocsNavProvider({ children }: { children: ReactNode }) {
     return () => window.removeEventListener("hashchange", close);
   }, [close]);
 
+  // The drawer markup is md:hidden, so if the viewport is resized or rotated
+  // past the md breakpoint while the drawer is open, the drawer (and its close
+  // button) disappears while body scroll stays locked. Close on crossing.
+  useEffect(() => {
+    const mql = window.matchMedia("(min-width: 768px)");
+    const handleChange = (event: MediaQueryListEvent) => {
+      if (event.matches) {
+        close();
+      }
+    };
+    mql.addEventListener("change", handleChange);
+    return () => mql.removeEventListener("change", handleChange);
+  }, [close]);
+
   // Cleanup on unmount
   useEffect(() => {
     return () => {
